--- Assets/Resources/Health Bar/README.md
@@ -1,1 +1,0 @@
-
