+++ Assets/Resources/TinyHealthSystem/Bars.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 819153730b8599e47878f66351cc233a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
