+++ Assets/Resources/TinyHealthSystem/Bars/Frame.psd.meta
