+++ Assets/Resources/TinyHealthSystem/Bars/HealthBar.anim.meta
