+++ Assets/Resources/TinyHealthSystem/Bars/HealthBar.controller.meta
