+++ Assets/Resources/TinyHealthSystem/Bars/HealthBar.psd.meta
