+++ Assets/Resources/TinyHealthSystem/Bars/ManaBar.anim.meta
