+++ Assets/Resources/TinyHealthSystem/Bars/ManaBar.controller.meta
