+++ Assets/Resources/TinyHealthSystem/Demo.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d0e5dabc8f8cec8478afee40d6539f4f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
