+++ Assets/Resources/TinyHealthSystem/Demo/Demo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 28d345066ba382b4cb3bc51aaf8dc67c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
