+++ Assets/Resources/TinyHealthSystem/Demo/DemoScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 730c50460715a514483f8f482da1a730
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
