+++ Assets/Resources/TinyHealthSystem/Demo/Fonts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3c60a02de96544145a3b21cebd3193c3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
