+++ Assets/Resources/TinyHealthSystem/Demo/Fonts/Pixelnauts.ttf.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 810eae3ac5e0d4444a7eb4576a501aa3
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontName: Pixelnauts
+  fontNames:
+  - Pixelnauts
+  fallbackFontReferences: []
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
