+++ Assets/Resources/TinyHealthSystem/Demo/PopupText.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a584f4f60d8c9964486a768f154d6a9a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
