+++ Assets/Resources/TinyHealthSystem/Globes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: de55435708db44e4783106480d2eb382
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
