+++ Assets/Resources/TinyHealthSystem/Globes/GlobeMask.png.meta
