+++ Assets/Resources/TinyHealthSystem/Globes/HealthGlobe.anim.meta
