+++ Assets/Resources/TinyHealthSystem/Globes/ManaGlobe.anim
