+++ Assets/Resources/TinyHealthSystem/Globes/ManaGlobe.controller.meta
