+++ Assets/Resources/TinyHealthSystem/Globes/ManaGlobe.psd.meta
