+++ Assets/Resources/TinyHealthSystem/HealthSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7875c39547040924ba157aab059c2394
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
