+++ Assets/Resources/TinyHealthSystem/TinyHealthSystem.pdf.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cebc2934d7c468f46a996af3d94343db
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
