+++ Assets/Resources/TinyHealthSystem/TinyHealthSystem.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f7162b84863baba4bafb2589864e6802
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
